python-client/requirements.txt
websocket-client==1.6.4
pyautogui==0.9.54
pillow>=9.0.0